Tidy up all of the files and folders
