+++ design-system/src/components/Article.js
@@ -1,10 +1,10 @@
 import React from "react";
 import styled from "styled-components";
 import { Route, Routes } from "react-router-dom";
-import Modal from "../features/Modal";
-import Tab from "../features/Tab";
-import Tag from "../features/Tag";
-import Toggle from "../features/Toggle";
+import ModalDemo from "../features/Modal/ModalDemo";
+import TabsDemo from "../features/Tabs/TabsDemo";
+import TagDemo from "../features/Tag/TagDemo";
+import ToggleDemo from "../features/Toggle/ToggleDemo";
 import AutoComplete from "../features/AutoComplete";
 import ClickToEdit from "../features/ClickToEdit";
 import ToggleButtons from "../features/ToggleButtons";
@@ -23,10 +23,10 @@ const Article = () => {
     <Container>
       <Routes>
         <Route path="/">Article</Route>
-        <Route path="/Modal" element={<Modal />} />
-        <Route path="/Tab" element={<Tab />} />
-        <Route path="/Tag" element={<Tag />} />
-        <Route path="/Toggle" element={<Toggle />} />
+        <Route path="/Modal" element={<ModalDemo />} />
+        <Route path="/Tabs" element={<TabsDemo />} />
+        <Route path="/Tag" element={<TagDemo />} />
+        <Route path="/Toggle" element={<ToggleDemo />} />
         <Route path="/AutoComplete" element={<AutoComplete />} />
         <Route path="/ClickToEdit" element={<ClickToEdit />} />
         <Route path="/ToggleButtons" element={<ToggleButtons />} />
